+++ trackerlogic.h
@@ -194,7 +194,7 @@ void   exerr(char *message);
    otherwise it is released in return_peers_for_torrent */
 size_t add_peer_to_torrent_and_return_peers(PROTO_FLAG proto, struct ot_workstruct *ws, size_t amount);
 size_t remove_peer_from_torrent(PROTO_FLAG proto, struct ot_workstruct *ws);
-size_t return_tcp_scrape_for_torrent(ot_hash const *hash_list, int amount, char *reply);
+size_t return_tcp_scrape_for_torrent(ot_hash const *hash_list, size_t amount, char *reply);
 size_t return_udp_scrape_for_torrent(ot_hash const hash, char *reply);
 void   add_torrent_from_saved_state(ot_hash const hash, ot_time base, size_t down_count);
 #ifdef _DEBUG_RANDOMTORRENTS
